+++ gen/com/intellij/idea/plugin/hybris/impex/psi/ImpexValueGroup.java
@@ -1,8 +1,4 @@
 /*
- * ----------------------------------------------------------------
- * --- WARNING: THIS FILE IS GENERATED AND WILL BE OVERWRITTEN! ---
- * ----------------------------------------------------------------
- *
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
  * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
  *
@@ -37,5 +33,8 @@ public interface ImpexValueGroup extends PsiElement {
 
   @Nullable
   ImpexValueLine getValueLine();
+
+  @Nullable
+  String computeValue();
 
 }