+++ gen/com/intellij/idea/plugin/hybris/impex/psi/impl/ImpexFullHeaderParameterImpl.java
@@ -1,8 +1,4 @@
 /*
- * ----------------------------------------------------------------
- * --- WARNING: THIS FILE IS GENERATED AND WILL BE OVERWRITTEN! ---
- * ----------------------------------------------------------------
- *
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
  * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
  *
@@ -30,6 +26,7 @@ import com.intellij.psi.util.PsiTreeUtil;
 import static com.intellij.idea.plugin.hybris.impex.psi.ImpexTypes.*;
 import com.intellij.extapi.psi.ASTWrapperPsiElement;
 import com.intellij.idea.plugin.hybris.impex.psi.*;
+import com.intellij.idea.plugin.hybris.impex.constants.modifier.AttributeModifier;
 
 public class ImpexFullHeaderParameterImpl extends ASTWrapperPsiElement implements ImpexFullHeaderParameter {
 
@@ -74,6 +71,12 @@ public class ImpexFullHeaderParameterImpl extends ASTWrapperPsiElement implement
   @Override
   public int getColumnNumber() {
     return ImpexPsiUtil.getColumnNumber(this);
+  }
+
+  @Override
+  @Nullable
+  public ImpexAttribute getAttribute(@NotNull AttributeModifier attributeModifier) {
+    return ImpexPsiUtil.getAttribute(this, attributeModifier);
   }
 
 }